PopUp component
Este aparece cada 15 seg y cuando es visible dura indefinidamente hasta que uno le da click en breath out
